refactor: 消除 L3 循环依赖并清理重复 Protocol 定义
依赖重定向（净 -183 行）：
- execution/runtime → orchestra 中转换为直接引用 observability（L6）
  影响：coordinator, auto_scene_recovery, governance_sync_runner,
        cleanup_executor, heartbeat, periodic_check_executor
- services → orchestra 的 orchestra_events_log_path/FailedGate 改为
  直接引用 observability/domain.failed_gate
  影响：orchestra_status_service, serve_status_service,
        scan_service, status_query_service
- queue_ordering 从 orchestra/ 迁移到 utils/（L6）
  status_query_service, queue_operations 同步更新

Protocol 清理：
- 删除 clients/protocols/services_protocols.py（196 行死代码）
- CheckServiceProtocol/TriggerableRoleDefinitionProtocol 合并入
  domain/protocols/dispatch_protocols.py（唯一真源）

效果：L3 内部循环从 12 个降至 8 个（消除 orchestra 作为中间节点）
