+++ src/vibe3/services/orchestra_status_service.py
@@ -15,7 +15,7 @@ from vibe3.clients import (
     parse_blocked_by,
 )
 from vibe3.models import IssueState, OrchestraConfig
-from vibe3.orchestra import orchestra_events_log_path
+from vibe3.observability import orchestra_events_log_path
 from vibe3.services.flow_reader import FlowReader
 from vibe3.services.label_service import LabelService
 from vibe3.services.orchestra_helpers import get_manager_usernames
