+++ src/vibe3/utils/queue_ordering.py
